refactor(utils): move Logger interface to @trezor/utils
The Logger interface lived in @trezor/transport-common, but logging is not a transport
concern — its sibling primitives (Log, LogMessage, LogWriter, LogsManager) already live in
@trezor/utils, and the Log class there already implements that exact shape. Move Logger next
to them, declare `Log implements Logger`, and update every consumer (transport-common's
AbstractTransport / AbstractApi / thp, and transport-bluetooth) to import it directly from
@trezor/utils.

Prerequisite for #28256 / #28266 (injected connect logger): lets @trezor/connect depend on a
general-purpose Logger type from @trezor/utils instead of from @trezor/transport-common.
